cognitive-core/base-agent/requirements.txt
google-adk
google-genai
fastapi
uvicorn[standard]
pydantic
google-cloud-aiplatform
python-multipart
pyyaml
